Move toxics and io_chan to their own packages
