+++ Dockerfile
@@ -1,7 +1,7 @@
-FROM golang:1.4 
+FROM golang:1.4
 
-ADD . /app
-RUN cd /app && GOPATH=/app/Godeps/_workspace go build -o toxiproxy
+ADD . /app/src/github.com/Shopify/toxiproxy
+RUN cd /app/src/github.com/Shopify/toxiproxy && GOPATH=/app/src/github.com/Shopify/toxiproxy/Godeps/_workspace:/app go build -o /app/toxiproxy ./cmd
 
 EXPOSE 8474
 ENTRYPOINT ["/app/toxiproxy"]
