+++ Makefile
@@ -9,31 +9,31 @@ COMBINED_GOPATH=$(GODEP_PATH):$(ORIGINAL_PATH)
 
 all: version deb linux darwin windows docker
 deb: $(DEB)
-darwin: tmp/build/toxiproxy-darwin-amd64 
+darwin: tmp/build/toxiproxy-darwin-amd64
 linux: tmp/build/toxiproxy-linux-amd64
 windows: tmp/build/toxiproxy-windows-amd64.exe
 
 build:
-	GOPATH=$(COMBINED_GOPATH) go build -o toxiproxy
+	GOPATH=$(COMBINED_GOPATH) go build -o toxiproxy ./cmd
 
 clean:
 	rm tmp/build/*
 	rm *.deb
 
 test:
-	GOMAXPROCS=4 GOPATH=$(COMBINED_GOPATH) go test -v -race
+	GOMAXPROCS=4 GOPATH=$(COMBINED_GOPATH) go test -v -race ./...
 
 version:
 	sed -i "s/Version = \"[^\"]*\"/Version = \"$(VERSION)\"/" version.go
 
 tmp/build/toxiproxy-linux-amd64:
-	GOOS=linux GOARCH=amd64 GOPATH=$(COMBINED_GOPATH) go build -o $(@)
+	GOOS=linux GOARCH=amd64 GOPATH=$(COMBINED_GOPATH) go build -o $(@) ./cmd
 
 tmp/build/toxiproxy-darwin-amd64:
-	GOOS=darwin GOARCH=amd64 GOPATH=$(COMBINED_GOPATH) go build -o $(@)
+	GOOS=darwin GOARCH=amd64 GOPATH=$(COMBINED_GOPATH) go build -o $(@) ./cmd
 
 tmp/build/toxiproxy-windows-amd64.exe:
-	GOOS=windows GOARCH=amd64 GOPATH=$(COMBINED_GOPATH) go build -o $(@)
+	GOOS=windows GOARCH=amd64 GOPATH=$(COMBINED_GOPATH) go build -o $(@) ./cmd
 
 docker:
 	docker build --tag="shopify/toxiproxy:$(VERSION)" .
