+++ proxy.go
@@ -1,4 +1,4 @@
-package main
+package toxiproxy
 
 import (
 	"errors"
@@ -28,8 +28,8 @@ type Proxy struct {
 
 	tomb        tomb.Tomb
 	connections ConnectionList
-	upToxics    *ToxicCollection
-	downToxics  *ToxicCollection
+	UpToxics    *ToxicCollection `json:"-"`
+	DownToxics  *ToxicCollection `json:"-"`
 }
 
 type ConnectionList struct {
@@ -52,8 +52,8 @@ func NewProxy() *Proxy {
 		started:     make(chan error),
 		connections: ConnectionList{list: make(map[string]net.Conn)},
 	}
-	proxy.upToxics = NewToxicCollection(proxy)
-	proxy.downToxics = NewToxicCollection(proxy)
+	proxy.UpToxics = NewToxicCollection(proxy)
+	proxy.DownToxics = NewToxicCollection(proxy)
 	return proxy
 }
 
@@ -177,8 +177,8 @@ func (proxy *Proxy) server() {
 		proxy.connections.list[name+"client"] = client
 		proxy.connections.list[name+"upstream"] = upstream
 		proxy.connections.Unlock()
-		proxy.upToxics.StartLink(name+"client", client, upstream)
-		proxy.downToxics.StartLink(name+"upstream", upstream, client)
+		proxy.UpToxics.StartLink(name+"client", client, upstream)
+		proxy.DownToxics.StartLink(name+"upstream", upstream, client)
 	}
 }
