+++ stream/io_chan_test.go
@@ -1,4 +1,4 @@
-package main
+package stream
 
 import (
 	"bytes"
