+++ toxic_collection.go
@@ -1,28 +1,30 @@
-package main
+package toxiproxy
 
 import (
 	"bytes"
 	"encoding/json"
 	"io"
 	"sync"
+
+	"github.com/Shopify/toxiproxy/toxics"
 )
 
 type ToxicCollection struct {
 	sync.Mutex
 
-	noop   *ToxicWrapper
+	noop   *toxics.ToxicWrapper
 	proxy  *Proxy
-	chain  []*ToxicWrapper
-	toxics []*ToxicWrapper
+	chain  []*toxics.ToxicWrapper
+	toxics []*toxics.ToxicWrapper
 	links  map[string]*ToxicLink
 }
 
 func NewToxicCollection(proxy *Proxy) *ToxicCollection {
 	collection := &ToxicCollection{
-		noop:   &ToxicWrapper{new(NoopToxic), "", "", 0},
+		noop:   &toxics.ToxicWrapper{new(toxics.NoopToxic), "", "", 0},
 		proxy:  proxy,
-		chain:  make([]*ToxicWrapper, 1, ToxicCount()+1),
-		toxics: make([]*ToxicWrapper, 0, ToxicCount()),
+		chain:  make([]*toxics.ToxicWrapper, 1, toxics.Count()+1),
+		toxics: make([]*toxics.ToxicWrapper, 0, toxics.Count()),
 		links:  make(map[string]*ToxicLink),
 	}
 	collection.chain[0] = collection.noop
@@ -40,7 +42,7 @@ func (c *ToxicCollection) ResetToxics() {
 	c.toxics = c.toxics[:0]
 }
 
-func (c *ToxicCollection) GetToxic(name string) Toxic {
+func (c *ToxicCollection) GetToxic(name string) toxics.Toxic {
 	c.Lock()
 	defer c.Unlock()
 
@@ -52,24 +54,24 @@ func (c *ToxicCollection) GetToxic(name string) Toxic {
 	return nil
 }
 
-func (c *ToxicCollection) GetToxicMap() map[string]Toxic {
+func (c *ToxicCollection) GetToxicMap() map[string]toxics.Toxic {
 	c.Lock()
 	defer c.Unlock()
 
-	result := make(map[string]Toxic)
+	result := make(map[string]toxics.Toxic)
 	for _, toxic := range c.toxics {
 		result[toxic.Name] = toxic
 	}
 	return result
 }
 
-func (c *ToxicCollection) AddToxicJson(data io.Reader) (Toxic, error) {
+func (c *ToxicCollection) AddToxicJson(data io.Reader) (toxics.Toxic, error) {
 	c.Lock()
 	defer c.Unlock()
 
 	var buffer bytes.Buffer
 
-	wrapper := new(ToxicWrapper)
+	wrapper := new(toxics.ToxicWrapper)
 	err := json.NewDecoder(io.TeeReader(data, &buffer)).Decode(wrapper)
 	if err != nil {
 		return nil, joinError(err, ErrBadRequestBody)
@@ -78,7 +80,7 @@ func (c *ToxicCollection) AddToxicJson(data io.Reader) (Toxic, error) {
 		wrapper.Name = wrapper.Type
 	}
 
-	wrapper.Toxic = NewToxic(wrapper.Type)
+	wrapper.Toxic = toxics.New(wrapper.Type)
 	if wrapper.Toxic == nil {
 		return nil, ErrInvalidToxicType
 	}
@@ -98,7 +100,7 @@ func (c *ToxicCollection) AddToxicJson(data io.Reader) (Toxic, error) {
 	return wrapper.Toxic, nil
 }
 
-func (c *ToxicCollection) UpdateToxicJson(name string, data io.Reader) (Toxic, error) {
+func (c *ToxicCollection) UpdateToxicJson(name string, data io.Reader) (toxics.Toxic, error) {
 	c.Lock()
 	defer c.Unlock()
 
@@ -147,7 +149,7 @@ func (c *ToxicCollection) RemoveLink(name string) {
 }
 
 // All following functions assume the lock is already grabbed
-func (c *ToxicCollection) chainAddToxic(toxic *ToxicWrapper) {
+func (c *ToxicCollection) chainAddToxic(toxic *toxics.ToxicWrapper) {
 	toxic.Index = len(c.chain)
 	c.chain = append(c.chain, toxic)
 
@@ -163,7 +165,7 @@ func (c *ToxicCollection) chainAddToxic(toxic *ToxicWrapper) {
 	group.Wait()
 }
 
-func (c *ToxicCollection) chainUpdateToxic(toxic *ToxicWrapper) {
+func (c *ToxicCollection) chainUpdateToxic(toxic *toxics.ToxicWrapper) {
 	c.chain[toxic.Index] = toxic
 
 	// Asynchronously add the toxic in each link
@@ -178,7 +180,7 @@ func (c *ToxicCollection) chainUpdateToxic(toxic *ToxicWrapper) {
 	group.Wait()
 }
 
-func (c *ToxicCollection) chainRemoveToxic(toxic *ToxicWrapper) {
+func (c *ToxicCollection) chainRemoveToxic(toxic *toxics.ToxicWrapper) {
 	c.chain = append(c.chain[:toxic.Index], c.chain[toxic.Index+1:]...)
 	for i := toxic.Index; i < len(c.chain); i++ {
 		c.chain[i].Index = i
