+++ toxics/slicer.go
@@ -1,8 +1,10 @@
-package main
+package toxics
 
 import (
 	"math/rand"
 	"time"
+
+	"github.com/Shopify/toxiproxy/stream"
 )
 
 // The SlicerToxic slices data into multiple smaller packets
@@ -47,26 +49,26 @@ func (t *SlicerToxic) chunk(start int, end int) []int {
 func (t *SlicerToxic) Pipe(stub *ToxicStub) {
 	for {
 		select {
-		case <-stub.interrupt:
+		case <-stub.Interrupt:
 			return
-		case c := <-stub.input:
+		case c := <-stub.Input:
 			if c == nil {
 				stub.Close()
 				return
 			}
 
-			chunks := t.chunk(0, len(c.data))
+			chunks := t.chunk(0, len(c.Data))
 			for i := 1; i < len(chunks); i += 2 {
-				stub.output <- &StreamChunk{
-					data:      c.data[chunks[i-1]:chunks[i]],
-					timestamp: c.timestamp,
+				stub.Output <- &stream.StreamChunk{
+					Data:      c.Data[chunks[i-1]:chunks[i]],
+					Timestamp: c.Timestamp,
 				}
 
 				select {
-				case <-stub.interrupt:
-					stub.output <- &StreamChunk{
-						data:      c.data[chunks[i]:],
-						timestamp: c.timestamp,
+				case <-stub.Interrupt:
+					stub.Output <- &stream.StreamChunk{
+						Data:      c.Data[chunks[i]:],
+						Timestamp: c.Timestamp,
 					}
 					return
 				case <-time.After(time.Duration(t.Delay) * time.Microsecond):
@@ -77,5 +79,5 @@ func (t *SlicerToxic) Pipe(stub *ToxicStub) {
 }
 
 func init() {
-	RegisterToxic("slicer", new(SlicerToxic))
+	Register("slicer", new(SlicerToxic))
 }
