+++ toxics/slow_close.go
@@ -1,4 +1,4 @@
-package main
+package toxics
 
 import "time"
 
@@ -11,24 +11,24 @@ type SlowCloseToxic struct {
 func (t *SlowCloseToxic) Pipe(stub *ToxicStub) {
 	for {
 		select {
-		case <-stub.interrupt:
+		case <-stub.Interrupt:
 			return
-		case c := <-stub.input:
+		case c := <-stub.Input:
 			if c == nil {
 				delay := time.Duration(t.Delay) * time.Millisecond
 				select {
 				case <-time.After(delay):
 					stub.Close()
 					return
-				case <-stub.interrupt:
+				case <-stub.Interrupt:
 					return
 				}
 			}
-			stub.output <- c
+			stub.Output <- c
 		}
 	}
 }
 
 func init() {
-	RegisterToxic("slow_close", new(SlowCloseToxic))
+	Register("slow_close", new(SlowCloseToxic))
 }
