+++ version.go
@@ -1,4 +1,4 @@
-package main
+package toxiproxy
 
 //go:generate make version
 var Version = "1.2.1"
